only user with role admin can access admin

+++ app/controllers/admin/product.go
@@ -4,6 +4,7 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/novanda1/sayurgo/app/models"
 	"github.com/novanda1/sayurgo/app/services"
+	"github.com/novanda1/sayurgo/pkg/utils"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 )
@@ -18,9 +19,15 @@ import (
 // @Param body body models.Product false "Product"
 // @Router /admin/products [post]
 func AdminCreateProduct(c *fiber.Ctx) error {
+	if !utils.AmIAdmin(c) {
+		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{
+			"success": false,
+			"error":   "401",
+		})
+	}
+
 	body := new(models.Product)
 	err := c.BodyParser(&body)
-
 	if err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
 			"success":  false,
@@ -62,10 +69,15 @@ func AdminCreateProduct(c *fiber.Ctx) error {
 // @Param id path int false "Product ID"
 // @Router /admin/products/{id} [put]
 func AdminUpdateProduct(c *fiber.Ctx) error {
+	if !utils.AmIAdmin(c) {
+		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{
+			"success": false,
+			"error":   "401",
+		})
+	}
+
 	paramID := c.Params("id")
-
 	id, err := primitive.ObjectIDFromHex(paramID)
-
 	if err != nil {
 		return c.Status(fiber.StatusOK).JSON(fiber.Map{
 			"success": false,
@@ -116,8 +128,14 @@ func AdminUpdateProduct(c *fiber.Ctx) error {
 // @Param id path int false "Product ID"
 // @Router /admin/products/{id} [delete]
 func AdminDeleteProduct(c *fiber.Ctx) error {
-	err := services.DeleteProduct(c)
+	if !utils.AmIAdmin(c) {
+		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{
+			"success": false,
+			"error":   "401",
+		})
+	}
 
+	err := services.DeleteProduct(c)
 	if err != nil {
 		if err == mongo.ErrNoDocuments {
 			return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
